Revert "Updated care_site_id and location_id from bigint to int"
This reverts commit 711b1a3dfa073115abbaadfc1601a7e5d40f562f.

+++ source/org.ohdsi.cdm.framework.common/DataReaders/v5/CareSiteDataReader.cs
@@ -164,7 +164,7 @@ namespace org.ohdsi.cdm.framework.common.DataReaders.v5
             switch (i)
             {
                 case 0:
-                    return typeof(int);
+                    return typeof(long);
 
                 case 1:
                     return typeof(string);
@@ -173,7 +173,7 @@ namespace org.ohdsi.cdm.framework.common.DataReaders.v5
                     return typeof(int);
 
                 case 3:
-                    return typeof(int);
+                    return typeof(long);
 
                 case 4:
                     return typeof(string);
